[Silver V] Title: 단어 정렬, Time: 660 ms, Memory: 26200 KB -BaekjoonHub

+++ 백준/Silver/1181. 단어 정렬/단어 정렬.java
@@ -9,15 +9,15 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int N = Integer.parseInt(st.nextToken());
-        String[] word = new String[N];
+        long N = Long.parseLong(st.nextToken());
+        String[] word = new String[(int) N];
 
         for (int i = 0; i < N; i++) {
             word[i] = br.readLine();
         }
         Arrays.stream(word)
                 .distinct()
-                .sorted(Comparator.comparingInt(String::length).thenComparing(String::compareTo))
+                .sorted(Comparator.comparingLong(String::length).thenComparing(String::compareTo))
                 .forEach(System.out::println);
     }
 }
